+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: gsteyn <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/16 06:01:35 by gsteyn            #+#    #+#             */
-/*   Updated: 2018/07/18 17:16:04 by gsteyn           ###   ########.fr       */
+/*   Updated: 2018/07/18 17:21:16 by gsteyn           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -99,6 +99,6 @@ void				ft_putnstr(char *str, size_t n);
 void				ft_putnstr_fd(char *str, int fd, size_t n);
 int					ft_is_num(char const *s);
 int					ft_abs(int in);
-size_t				ft_2pow(int pow);
+unsigned int			ft_2pow(int pow);
 
 #endif
